Boot: Switch to 32-bit Protected Mode, initialize segment registers, load and run a simple kernel, which prints "X" to the screen.
I've spread the boot loader code for 2 sectors. The 1st sector, whicn
BIOS loads and executes, contains basic print string and hex value and
disk load procedures. It loads the rest of the boot loader code and the
kernel. It places the kernel code to the address 0x1000 and calls it.

Next: Let's explore the kernel further ;)